requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.3.0
PyMySQL==0.9.2
sqlacodegen==2.0.1
